Adds priming for the second i2c bus (0)
The second revision of the Raspberry pi links the
i2c bus 0 to the s5 header on boot for the pi camera
support.
To re-route the bus and make it communicate
over the p5 header, it's required to set gpio pins
0 and 1 (representing the s5 header i2c pins) to std
inputs. This releases their alternate function and
when followed by priming the p5 gpio pins the bus
should now be correctly routed.